Update simple class names to avoid conflicts
