+++ NPCSk/src/main/java/space/arim/npcsk/syntax/expr/ExprLastCreatedNPCSk.java
@@ -1,0 +1,73 @@
+/* 
+ * NPCSk, a robust Skript NPC addon
+ * Copyright © 2019 Anand Beh <https://www.arim.space>
+ * 
+ * NPCSk is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * NPCSk is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with NPCSk. If not, see <https://www.gnu.org/licenses/>
+ * and navigate to version 3 of the GNU General Public License.
+ */
+package space.arim.npcsk.syntax.expr;
+
+import org.eclipse.jdt.annotation.Nullable;
+
+import org.bukkit.event.Event;
+
+import space.arim.npcsk.NPCSk;
+
+import ch.njol.skript.Skript;
+import ch.njol.skript.doc.Description;
+import ch.njol.skript.doc.Name;
+import ch.njol.skript.doc.Since;
+import ch.njol.skript.lang.Expression;
+import ch.njol.skript.lang.ExpressionType;
+import ch.njol.skript.lang.SkriptParser.ParseResult;
+import ch.njol.skript.lang.util.SimpleExpression;
+import ch.njol.util.Kleenean;
+
+@Name("NPCSk Last Created NPC")
+@Description("The id of the last created NPC.")
+@Since("0.6.0")
+public class ExprLastCreatedNPCSk extends SimpleExpression<String> {
+
+	static {
+		Skript.registerExpression(ExprLastCreatedNPCSk.class, String.class, ExpressionType.SIMPLE, "[npcsk] last created npc");
+	}
+	
+	@Override
+	public Class<? extends String> getReturnType() {
+		return String.class;
+	}
+
+	@Override
+	public boolean isSingle() {
+		return true;
+	}
+
+	@Override
+	public boolean init(Expression<?>[] arg0, int arg1, Kleenean arg2, ParseResult arg3) {
+		return true;
+	}
+
+	@Override
+	public String toString(@Nullable Event evt, boolean debug) {
+		return "npcsk last created npc";
+	}
+
+	@Override
+	@Nullable
+	protected String[] get(Event arg0) {
+		String id = NPCSk.npcs().getLatestId();
+		return id != null ? new String[] {id} : null;
+	}
+
+}
