+++ NPCSk/src/main/java/space/arim/npcsk/syntax/expr/ExprVisibilityNPCSk.java
@@ -18,12 +18,14 @@
  */
 package space.arim.npcsk.syntax.expr;
 
-import org.bukkit.Location;
+import org.bukkit.entity.Player;
 import org.bukkit.event.Event;
 import org.eclipse.jdt.annotation.Nullable;
 
 import ch.njol.skript.Skript;
+import ch.njol.skript.classes.Changer.ChangeMode;
 import ch.njol.skript.doc.Description;
+import ch.njol.skript.doc.Examples;
 import ch.njol.skript.doc.Name;
 import ch.njol.skript.doc.Since;
 import ch.njol.skript.lang.Expression;
@@ -31,45 +33,63 @@ import ch.njol.skript.lang.ExpressionType;
 import ch.njol.skript.lang.SkriptParser.ParseResult;
 import ch.njol.skript.lang.util.SimpleExpression;
 import ch.njol.util.Kleenean;
+import ch.njol.util.coll.CollectionUtils;
+
 import space.arim.npcsk.NPCSk;
 
-@Name("NPCSk NPC Location")
-@Description("The location of a NPC; this cannot be changed. The only way to move an NPC is to delete and recreate it.")
-@Since("0.6.0")
-public class ExprLocation extends SimpleExpression<Location> {
+@Name("NPCSk Visibility")
+@Description("Whether a NPC is visible for a player. NPCs may be hidden or shown for specific players")
+@Examples({"on npc interact:", "\tset npc visibility of npc-event-npc for npc-event-player to true"})
+@Since("0.7.0")
+public class ExprVisibilityNPCSk extends SimpleExpression<Boolean> {
 	
 	private Expression<String> id;
+	private Expression<Player> target;
 	
 	static {
-		Skript.registerExpression(ExprLocation.class, Location.class, ExpressionType.PROPERTY, "[npcsk] npc location of [id] %string%");
+		Skript.registerExpression(ExprVisibilityNPCSk.class, Boolean.class, ExpressionType.COMBINED, "[npcsk] npc visibility of %string% for %player%");
 	}
-
+	
 	@Override
-	public Class<? extends Location> getReturnType() {
-		return Location.class;
+	public Class<? extends Boolean> getReturnType() {
+		return Boolean.class;
 	}
-
+	
 	@Override
 	public boolean isSingle() {
 		return true;
 	}
-
+	
 	@SuppressWarnings("unchecked")
 	@Override
 	public boolean init(Expression<?>[] exprs, int arg1, Kleenean arg2, ParseResult arg3) {
 		id = (Expression<String>) exprs[0];
+		target = (Expression<Player>) exprs[1];
 		return true;
 	}
-
+	
 	@Override
 	public String toString(@Nullable Event evt, boolean debug) {
-		return "npcsk npc location of npc with id " + id.toString(evt, debug);
+		return "npcsk npc visibility of " + id.toString(evt, debug) + " for " + target.toString(evt, debug);
 	}
-
+	
 	@Override
 	@Nullable
-	protected Location[] get(Event evt) {
-		return NPCSk.npcs().hasNpc(id.getSingle(evt)) ? new Location[] {NPCSk.npcs().getLocation(id.getSingle(evt))} : null;
+	protected Boolean[] get(Event evt) {
+		return new Boolean[] {NPCSk.npcs().isShown(id.getSingle(evt), target.getSingle(evt))};
+	}
+	
+	@Override
+	public Class<?>[] acceptChange(ChangeMode mode) {
+		if (mode == ChangeMode.SET || mode == ChangeMode.RESET) {
+			return CollectionUtils.array(Boolean.class);
+		}
+		return null;
+	}
+	
+	@Override
+	public void change(Event evt, Object[] delta, ChangeMode mode) {
+		NPCSk.npcs().setShown(id.getSingle(evt), target.getSingle(evt), mode == ChangeMode.SET && (Boolean) delta[0]);
 	}
 	
 }
